requirements.txt
telethon==1.35.0
yarl==1.9.4
loguru==0.7.2
pysocks==1.7.1
python-telegram-bot==20.6
python-dotenv==1.0.1

